% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{console_appender}
\alias{console_appender}
\alias{basic_appenders}
\alias{file_appender}
\title{Append Logs to a File or the Console}
\usage{
console_appender(layout = default_log_layout)

file_appender(file, append = TRUE, layout = default_log_layout)
}
\arguments{
\item{layout}{A layout function taking a \code{level} parameter and
additional arguments corresponding to the message. See
\code{\link{default_log_layout}}.}

\item{file}{The file to write messages to.}

\item{append}{When \code{TRUE}, the file is not truncated when opening for
the first time.}
}
\description{
The most basic appenders log messages to the console or to a file.
}
