% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.logger}
\alias{create.logger}
\title{Creates a logger object.}
\usage{
create.logger(logfile = "logfile.log", level = "FATAL", logformat = NULL)
}
\arguments{
\item{logfile}{The full pathname of the file you want log messages to be
written to.}

\item{level}{The level at which the logger is initialized.  Will be coerced
using \code{\link{as.loglevel}}.}

\item{logformat}{The format string used when writing messages to the log
file.}
}
\description{
Creates a logger object.
}
\examples{
library('log4r')

logger <- create.logger(logfile = 'debugging.log', level = "DEBUG")
}
\seealso{
\code{\link{loglevel}}, \code{\link{level.logger}}
}

