% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{default_log_layout}
\alias{default_log_layout}
\title{Default Log Message Layout}
\usage{
default_log_layout(level, ...)
}
\arguments{
\item{level}{The message level.}

\item{...}{Additional message arguments.}
}
\description{
Default Log Message Layout
}
