% Generated by roxygen2 (4.0.2): do not edit by hand
\name{error}
\alias{error}
\title{Write messages to logs at the error priority level.}
\usage{
error(logger, message)
}
\arguments{
\item{logger}{An object of class 'logger'.}

\item{message}{A string to be printed to the log with priority level
\code{log4r:::ERROR}.}
}
\description{
Write messages to logs at the error priority level.
}
\examples{
library('log4r')

logger <- create.logger(logfile = 'error.log', level = log4r:::ERROR)

error(logger, 'Generating an error message.')
}

