% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{gelf_log_layout}
\alias{gelf_log_layout}
\title{Layout Messages According to the Graylog Extended Log Format (GELF)}
\usage{
gelf_log_layout(host)
}
\arguments{
\item{host}{A string identifying the application.}
}
\description{
GELF is a JSON-based log format designed to overcome some of the limitations
of Syslog. Originally created for use in the open-source Graylog project, it
is now supported by many open and proprietary logging systems.

GELF-formatted logs can be delivered over TCP or HTTP.

Note that JSON encoding is currently expensive, on the order of 0.25 ms, and
requires the \code{jsonlite} package.
}
\seealso{
\href{The GELF Specification}{http://docs.graylog.org/en/3.0/pages/gelf.html}.
}
