% Generated by roxygen2 (4.0.2): do not edit by hand
\name{level<-.logger}
\alias{level<-}
\alias{level<-.logger}
\title{Set the priority level for a logger object.}
\usage{
\method{level}{logger}(x) <- value
}
\arguments{
\item{x}{An object of class logger.}

\item{value}{A level input, specified using the symbolic constants
\code{log4r:::DEBUG}, \code{log4r:::INFO}, \code{log4r:::WARN},
\code{log4r:::ERROR} and \code{log4r:::FATAL}.}
}
\description{
The priority level can be one of:
\code{log4r:::DEBUG}, \code{log4r:::INFO}, \code{log4r:::WARN},
\code{log4r:::ERROR} or \code{log4r:::FATAL}
}
\examples{
library('log4r')

logger <- create.logger(logfile = 'debugging.log', level = log4r:::DEBUG)

level(logger) <- log4r:::FATAL
}

