% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{level}
\alias{level}
\alias{level<-}
\alias{level.logger}
\alias{level<-.logger}
\alias{available.loglevels}
\title{Set the logging threshold level for a logger dynamically}
\usage{
level(x)

level(x) <- value

\method{level}{logger}(x)

\method{level}{logger}(x) <- value

available.loglevels()
}
\arguments{
\item{x}{An object of class \code{"logger"}.}

\item{value}{One of \code{"DEBUG"}, \code{"INFO"}, \code{"WARN"}, \code{"ERROR"}, or \code{"FATAL"}.}
}
\description{
It can sometimes be useful to change the logging threshold level at runtime.
The \code{\link[=level]{level()}} accessor allows doing so.
}
\examples{
lgr <- logger()
level(lgr) # Prints "INFO".
info(lgr, "This message is shown.")
level(lgr) <- "FATAL"
info(lgr, "This message is now suppressed.")
}
