% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{create.logger}
\alias{create.logger}
\alias{logfile}
\alias{logfile<-}
\alias{logfile.logger}
\alias{logfile<-.logger}
\alias{logformat}
\alias{logformat<-}
\alias{is.loglevel}
\alias{loglevel}
\alias{as.loglevel}
\alias{print.loglevel}
\alias{as.numeric.loglevel}
\alias{as.character.loglevel}
\alias{verbosity}
\alias{levellog}
\alias{debug}
\alias{info}
\alias{warn}
\alias{error}
\alias{fatal}
\title{Deprecated logger functions}
\usage{
create.logger(logfile = "logfile.log", level = "FATAL", logformat = NULL)

logfile(x)

logfile(x) <- value

\method{logfile}{logger}(x)

\method{logfile}{logger}(x) <- value

logformat(x)

logformat(x) <- value

is.loglevel(x, ...)

loglevel(i)

as.loglevel(i)

\method{print}{loglevel}(x, ...)

\method{as.numeric}{loglevel}(x, ...)

\method{as.character}{loglevel}(x, ...)

verbosity(v)

levellog(logger, level, ...)

debug(logger, ...)

info(logger, ...)

warn(logger, ...)

error(logger, ...)

fatal(logger, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\itemize{
\item \code{\link[=create.logger]{create.logger()}} and \code{\link[=logfile]{logfile()}} are deprecated in favour of \code{\link[=logger]{logger()}}.
They issue deprecation warnings when used.
\item \code{\link[=debug]{debug()}}, \code{\link[=info]{info()}}, \code{\link[=warn]{warn()}}, \code{\link[=error]{error()}}, and \code{\link[=fatal]{fatal()}} are deprecated in
favour of \code{\link[=log_debug]{log_debug()}}, \code{\link[=log_info]{log_info()}}, \code{\link[=log_warn]{log_warn()}}, \code{\link[=log_error]{log_error()}}, and
\code{\link[=log_fatal]{log_fatal()}}, respectively. For performance reasons they do not yet issue
deprecation warnings when used.
\item \code{\link[=levellog]{levellog()}} is deprecated in favour of \code{\link[=log_at]{log_at()}}. It issues a deprecation
warning when used.
\item \code{\link[=logformat]{logformat()}} is incompatible with \strong{\link[=layouts]{Layouts}} and has been
nonfunctional for many years. It issues a deprecation error when used.
\item \code{\link[=is.loglevel]{is.loglevel()}}, \code{\link[=as.loglevel]{as.loglevel()}}, its alias \code{\link[=loglevel]{loglevel()}}, and S3 generics
for the \code{"loglevel"} class are now considered an implementation detail and
are no longer part of the public API. They issue deprecation warnings when
used.
\item \code{\link[=verbosity]{verbosity()}} is similar, in that there is no longer a stable mapping
between priority integers and levels. It issues a deprecation warning when
used.
}
}
\keyword{internal}
