% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log4r-package.R
\docType{package}
\name{log4r-package}
\alias{log4r}
\alias{log4r-package}
\title{log4r: A Fast and Lightweight Logging System for R, Based on 'log4j'}
\description{
The log4r package is meant to provide a fast, lightweight, object-oriented approach to logging in R based on the widely-emulated 'log4j' system and etymology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnmyleswhite/log4r}
  \item Report bugs at \url{https://github.com/johnmyleswhite/log4r/issues}
}

}
\author{
\strong{Maintainer}: Aaron Jacobs \email{atheriel@gmail.com}

Authors:
\itemize{
  \item John Myles White [copyright holder]
}

Other contributors:
\itemize{
  \item Kenton White [contributor]
  \item Kirill Müller \email{krlmlr+r@mailbox.org} [contributor]
}

}
\keyword{internal}
