% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfuncs.R
\name{log_at}
\alias{log_at}
\alias{log_debug}
\alias{log_info}
\alias{log_warn}
\alias{log_error}
\alias{log_fatal}
\title{Write logs at a given level}
\usage{
log_at(logger, level, ...)

log_debug(logger, ...)

log_info(logger, ...)

log_warn(logger, ...)

log_error(logger, ...)

log_fatal(logger, ...)
}
\arguments{
\item{logger}{An object of class \code{"logger"}.}

\item{level}{The desired severity, one of \code{"DEBUG"}, \code{"INFO"}, \code{"WARN"},
\code{"ERROR"}, or \code{"FATAL"}. Messages with a lower severity than the logger
threshold will be discarded.}

\item{...}{One or more items to log.}
}
\description{
Write logs at a given level
}
\examples{
logger <- logger()

log_at(logger, "WARN", "First warning from our code")
log_debug(logger, "Debugging our code")
log_info(logger, "Information about our code")
log_warn(logger, "Another warning from our code")
log_error(logger, "An error from our code")
log_fatal(logger, "I'm outta here")
}
