% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logfile<-.logger}
\alias{logfile<-}
\alias{logfile<-.logger}
\title{Set the logfile for a logger object.}
\usage{
\method{logfile}{logger}(x) <- value
}
\arguments{
\item{x}{An object of class logger.}

\item{value}{The path name of a file to be used for logging.}
}
\description{
Set the logfile for a logger object. Must be a valid path in an already
existing directory.
}
\examples{
library('log4r')

logger <- create.logger()

logfile(logger) <- 'debug.log'

debug(logger, 'A Debugging Message')
}

