% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logfile.logger}
\alias{logfile}
\alias{logfile.logger}
\title{Get the logfile for a logger object.}
\usage{
\method{logfile}{logger}(x)
}
\arguments{
\item{x}{An object of class logger.}
}
\description{
Get the logfile for a logger object.
}
\examples{
library('log4r')

logger <- create.logger()

print(logfile(logger))
}

