% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logformat<-.logger}
\alias{logformat<-}
\alias{logformat<-.logger}
\title{Set the format string for a logger object.}
\usage{
\method{logformat}{logger}(x) <- value
}
\arguments{
\item{x}{An object of class logger.}

\item{value}{A string containing a proper format string.}
}
\description{
Set the format string for a logger object.
}
\examples{
library('log4r')

logger <- create.logger(logfile = 'debugging.log', level = log4r:::DEBUG)

logformat(logger) <- 'FORMAT STRING'
}

