% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logformat.logger}
\alias{logformat}
\alias{logformat.logger}
\title{Get the format string for a logger object.}
\usage{
\method{logformat}{logger}(x)
}
\arguments{
\item{x}{An object of class logger.}
}
\description{
Get the format string for a logger object.
}
\examples{
library('log4r')

logger <- create.logger(logfile = 'debugging.log', level = log4r:::DEBUG)

print(logformat(logger))
}

