% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{syslog_appender}
\alias{syslog_appender}
\title{Send logs to the local syslog}
\usage{
syslog_appender(identifier, layout = bare_log_layout(), ...)
}
\arguments{
\item{identifier}{A string identifying the application.}

\item{layout}{A layout function taking a \code{level} parameter and
additional arguments corresponding to the message.}

\item{...}{Further arguments passed on to \code{\link[rsyslog:syslog]{rsyslog::open_syslog()}}.}
}
\description{
Send messages to the local syslog. Requires the \code{rsyslog} package.
}
\seealso{
\link{appenders} for more information on Appenders.
}
