% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{tcp_appender}
\alias{tcp_appender}
\title{Send logs over TCP}
\usage{
tcp_appender(
  host,
  port,
  layout = default_log_layout(),
  timeout = getOption("timeout")
)
}
\arguments{
\item{host}{Hostname for the socket connection.}

\item{port}{Port number for the socket connection.}

\item{layout}{A layout function taking a \code{level} parameter and
additional arguments corresponding to the message.}

\item{timeout}{Timeout for the connection.}
}
\description{
Append messages to arbitrary TCP destinations.
}
\seealso{
\link{appenders} for more information on Appenders, and
\code{\link[base:connections]{base::socketConnection()}} for the underlying connection object
used by \code{\link[=tcp_appender]{tcp_appender()}}.
}
