% Generated by roxygen2 (4.0.2): do not edit by hand
\name{warn}
\alias{warn}
\title{Write messages to logs at the warn priority level.}
\usage{
warn(logger, message)
}
\arguments{
\item{logger}{An object of class 'logger'.}

\item{message}{A string to be printed to the log with priority level
\code{log4r:::WARN}.}
}
\description{
Write messages to logs at the warn priority level.
}
\examples{
library('log4r')

logger <- create.logger(logfile = 'warnings.log', level = log4r:::WARN)

warn(logger, 'Generating an warning message.')
}

