% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.message}
\alias{write.message}
\title{A hidden function for handling the writing of logging messages.}
\usage{
write.message(logger, message)
}
\arguments{
\item{logger}{An object of class logger.}

\item{message}{A message to be written to the log file. The current
timestamp will be appended to this message.}
}
\description{
A hidden function for handling the writing of logging messages.
}

