% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fatal}
\alias{fatal}
\title{Write messages to logs at the fatal priority level.}
\usage{
fatal(logger, message)
}
\arguments{
\item{logger}{An object of class 'logger'.}

\item{message}{A string to be printed to the log with priority level
\code{log4r:::FATAL}.}
}
\description{
Write messages to logs at the fatal priority level.
}
\examples{
library('log4r')

logger <- create.logger(logfile = 'fatal_error.log', level = log4r:::FATAL)

fatal(logger, 'Generating a fatal error message.')
}

