% Generated by roxygen2 (4.0.2): do not edit by hand
\name{level.logger}
\alias{level}
\alias{level.logger}
\title{Get the priority level for a logger object.}
\usage{
\method{level}{logger}(x)
}
\arguments{
\item{x}{An object of class logger.}
}
\description{
The priority level can be one of:
\code{log4r:::DEBUG}, \code{log4r:::INFO}, \code{log4r:::WARN},
\code{log4r:::ERROR} or \code{log4r:::FATAL}
}
\examples{
library('log4r')

logger <- create.logger(logfile = 'debugging.log', level = log4r:::DEBUG)

print(level(logger))
}

