% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile.logger.R
\name{logfile}
\alias{logfile}
\alias{logfile<-}
\alias{logfile.logger}
\alias{logfile<-.logger}
\title{Get or set the logfile for a logger object.}
\usage{
logfile(x)

logfile(x) <- value

\method{logfile}{logger}(x)

\method{logfile}{logger}(x) <- value
}
\arguments{
\item{x}{An object of class logger.}

\item{value}{The path name of a file to be used for logging. Must be a valid
path in an already existing directory}
}
\description{
Get or set the logfile for a logger object.
}
\examples{

library('log4r')

logger <- create.logger()
print(logfile(logger))
logfile(logger) <- 'debug.log'
debug(logger, 'A Debugging Message')
}
